requirements.txt
streamlit
chromadb
google-generativeai
python-dotenv